Use existing RuntimeType class in common
