+++ pom.xml
@@ -24,7 +24,6 @@
         <agroal.version>2.8</agroal.version>
         <spring.version>6.2.11</spring.version>
         <spring-boot.version>3.5.5</spring-boot.version>
-        <spring-framework.version>6.2.11</spring-framework.version>
         <quarkus.version>3.26.4</quarkus.version>
         <microprofile.version>3.1</microprofile.version>
         <smallrye.version>3.14.0</smallrye.version>
